more unit test cleaning/ fixing
